✨ :: [#434] ArtistDomain 모듈 작업 완료 및 레거시 제거
