+++ Projects/Domains/ArtistDomain/Interface/DataSource/RemoteArtistDataSource.swift
@@ -1,17 +1,15 @@
 //
-//  ArtistRepository.swift
-//  DomainModule
+//  RemoteArtistDataSource.swift
+//  NetworkModuleTests
 //
 //  Created by KTH on 2023/02/08.
 //  Copyright © 2023 yongbeomkwak. All rights reserved.
 //
 
-import Foundation
 import RxSwift
-import DataMappingModule
-import ErrorModule
+import Foundation
 
-public protocol ArtistRepository {
+public protocol RemoteArtistDataSource {
     func fetchArtistList() -> Single<[ArtistListEntity]>
     func fetchArtistSongList(id: String, sort: ArtistSongSortType, page: Int) -> Single<[ArtistSongListEntity]>
 }
