+++ Projects/Domains/ArtistDomain/Interface/Entity/ArtistListEntity.swift
