+++ Projects/Domains/ArtistDomain/Interface/Repository/ArtistRepository.swift
@@ -1,18 +1,15 @@
 //
-//  RemoteArtistDataSource.swift
-//  NetworkModuleTests
+//  ArtistRepository.swift
+//  DomainModule
 //
 //  Created by KTH on 2023/02/08.
 //  Copyright © 2023 yongbeomkwak. All rights reserved.
 //
 
-import DataMappingModule
-import ErrorModule
-import DomainModule
-import RxSwift
 import Foundation
+import RxSwift
 
-public protocol RemoteArtistDataSource {
+public protocol ArtistRepository {
     func fetchArtistList() -> Single<[ArtistListEntity]>
     func fetchArtistSongList(id: String, sort: ArtistSongSortType, page: Int) -> Single<[ArtistSongListEntity]>
 }
