--- Projects/Services/NetworkModule/Sources/Artist/DataTransfer/FetchArtistSongListTransfer.swift
@@ -1,28 +1,0 @@
-//
-//  FetchArtistSongListTransfer.swift
-//  NetworkModule
-//
-//  Created by KTH on 2023/02/10.
-//  Copyright © 2023 yongbeomkwak. All rights reserved.
-//
-
-import Foundation
-import DataMappingModule
-import DomainModule
-import Utility
-
-public extension ArtistSongListResponseDTO {
-    func toDomain() -> ArtistSongListEntity {
-        ArtistSongListEntity(
-            ID: ID,
-            title: title,
-            artist: artist,
-            remix: remix,
-            reaction: reaction,
-            date: date.changeDateFormat(origin: "yyMMdd", result: "yyyy.MM.dd"),
-            views: total?.views ?? 0,
-            last: total?.last ?? 0,
-            isSelected: false
-        )
-    }
-}
